Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch import-data-admin
Changes to be committed:
	modified:   app/Http/Controllers/WEB/Admin/AdminController.php
	modified:   app/Http/Controllers/WEB/Admin/DosenController.php
	modified:   app/Http/Controllers/WEB/Admin/MahasiswaController.php
	modified:   app/Http/Controllers/WEB/Admin/MataKuliahController.php
	new file:   app/Http/Controllers/WEB/DashboardController.php
	deleted:    app/Http/Controllers/WEB/Dosen/DosenController.php
	new file:   app/Imports/DosenImport.php
	modified:   app/Imports/MahasiswaImport.php
	modified:   app/Imports/MatkulImport.php
	deleted:    resources/views/admin/kelas/index.blade.php
	deleted:    resources/views/admin/pengguna/adminandstaff/index.blade.php
	deleted:    resources/views/admin/pengguna/adminandstaff/table.blade.php
	deleted:    resources/views/admin/pengguna/dosen/index.blade.php
	deleted:    resources/views/admin/pengguna/dosen/table.blade.php
	renamed:    resources/views/admin/dashboard/index.blade.php -> resources/views/dashboard/index.blade.php
	deleted:    resources/views/dosen/index.blade.php
	new file:   resources/views/pageAdmin/kelas/index.blade.php
	new file:   resources/views/pageAdmin/kelas/modal/import.blade.php
	new file:   resources/views/pageAdmin/kelas/modal/tambah.blade.php
	renamed:    resources/views/admin/kelas/table.blade.php -> resources/views/pageAdmin/kelas/table.blade.php
	renamed:    resources/views/admin/matakuliah/index.blade.php -> resources/views/pageAdmin/matakuliah/index.blade.php
	renamed:    resources/views/admin/matakuliah/modal/edit.blade.php -> resources/views/pageAdmin/matakuliah/modal/edit.blade.php
	renamed:    resources/views/admin/matakuliah/modal/import.blade.php -> resources/views/pageAdmin/matakuliah/modal/import.blade.php
	renamed:    resources/views/admin/matakuliah/modal/tambah.blade.php -> resources/views/pageAdmin/matakuliah/modal/tambah.blade.php
	renamed:    resources/views/admin/matakuliah/table.blade.php -> resources/views/pageAdmin/matakuliah/table.blade.php
	new file:   resources/views/pageAdmin/pengguna/adminandstaff/index.blade.php
	new file:   resources/views/pageAdmin/pengguna/adminandstaff/modal/detail.blade.php
	new file:   resources/views/pageAdmin/pengguna/adminandstaff/modal/edit.blade.php
	new file:   resources/views/pageAdmin/pengguna/adminandstaff/modal/tambah.blade.php
	new file:   resources/views/pageAdmin/pengguna/adminandstaff/table.blade.php
	new file:   resources/views/pageAdmin/pengguna/dosen/index.blade.php
	new file:   resources/views/pageAdmin/pengguna/dosen/modal/edit.blade.php
	new file:   resources/views/pageAdmin/pengguna/dosen/modal/import.blade.php
	new file:   resources/views/pageAdmin/pengguna/dosen/modal/tambah.blade.php
	new file:   resources/views/pageAdmin/pengguna/dosen/table.blade.php
	renamed:    resources/views/admin/pengguna/mahasiswa/index.blade.php -> resources/views/pageAdmin/pengguna/mahasiswa/index.blade.php
	renamed:    resources/views/admin/pengguna/mahasiswa/modal/detail.blade.php -> resources/views/pageAdmin/pengguna/mahasiswa/modal/detail.blade.php
	renamed:    resources/views/admin/pengguna/mahasiswa/modal/edit.blade.php -> resources/views/pageAdmin/pengguna/mahasiswa/modal/edit.blade.php
	renamed:    resources/views/admin/pengguna/mahasiswa/modal/import.blade.php -> resources/views/pageAdmin/pengguna/mahasiswa/modal/import.blade.php
	renamed:    resources/views/admin/pengguna/mahasiswa/modal/tambah.blade.php -> resources/views/pageAdmin/pengguna/mahasiswa/modal/tambah.blade.php
	renamed:    resources/views/admin/pengguna/mahasiswa/table.blade.php -> resources/views/pageAdmin/pengguna/mahasiswa/table.blade.php
	renamed:    resources/views/admin/barang/actions.blade.php -> resources/views/pageStaff/barang/actions.blade.php
	renamed:    resources/views/admin/barang/index.blade.php -> resources/views/pageStaff/barang/index.blade.php
	renamed:    resources/views/admin/barang/table.blade.php -> resources/views/pageStaff/barang/table.blade.php
	renamed:    resources/views/admin/kategori/index.blade.php -> resources/views/pageStaff/kategori/index.blade.php
	renamed:    resources/views/admin/kategori/table.blade.php -> resources/views/pageStaff/kategori/table.blade.php
	renamed:    resources/views/admin/peminjaman/index.blade.php -> resources/views/pageStaff/peminjaman/index.blade.php
	renamed:    resources/views/admin/ruangan/index.blade.php -> resources/views/pageStaff/ruangan/index.blade.php
	renamed:    resources/views/admin/ruangan/modal/edit.blade.php -> resources/views/pageStaff/ruangan/modal/edit.blade.php
	renamed:    resources/views/admin/ruangan/modal/import.blade.php -> resources/views/pageStaff/ruangan/modal/import.blade.php
	renamed:    resources/views/admin/ruangan/modal/tambah.blade.php -> resources/views/pageStaff/ruangan/modal/tambah.blade.php
	renamed:    resources/views/admin/ruangan/table.blade.php -> resources/views/pageStaff/ruangan/table.blade.php
	renamed:    resources/views/admin/satuan/index.blade.php -> resources/views/pageStaff/satuan/index.blade.php
	renamed:    resources/views/admin/satuan/table.blade.php -> resources/views/pageStaff/satuan/table.blade.php
	modified:   routes/web.php
