--- resources/views/admin/pengguna/adminandstaff/index.blade.php
@@ -1,210 +1,0 @@
-@extends('index')
-@section('content')
-    <div class="p-4 mt-3 sm:ml-64">
-        <div class="space-y-4 rounded-lg mt-14">
-            @if (session('success'))
-                <script>
-                    Swal.fire({
-                        title: "Success",
-                        text: "{{ session('success') }}",
-                        icon: "success",
-                        confirmButtonColor: "#3085d6",
-                    });
-                </script>
-            @endif
-            <div class="p-4 bg-white rounded-lg shadow-lg">
-                <div class="flex items-center justify-between">
-                    <div>
-                        <h3 class="text-2xl font-semibold text-green-500">Data Admin dan Staff</h3>
-                    </div>
-                    <div>
-                        <button data-modal-target="barang" data-modal-toggle="barang"
-                            class="px-3 py-2 text-white bg-green-500 rounded hover:bg-green-800"><i
-                                class="fa-solid fa-plus"></i>
-                        </button>
-
-                        {{-- MODAL TAMBAH ADMIN DAN STAFF --}}
-                        <div id="barang" tabindex="-1" aria-hidden="true"
-                            class="hidden overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 z-50 justify-center items-center w-full md:inset-0 h-[calc(100%-1rem)] max-h-full">
-                            <div class="relative w-full max-w-2xl max-h-full p-4">
-                                <!-- Modal content -->
-                                <div class="relative bg-white rounded-lg shadow dark:bg-gray-700">
-                                    <!-- Modal header -->
-                                    <div
-                                        class="flex items-center justify-between p-4 border-b rounded-t md:p-5 dark:border-gray-600">
-                                        <h3 class="text-xl font-semibold text-gray-900 dark:text-white">
-                                            Tambah Data Admin dan Staff
-                                        </h3>
-                                        <button type="button"
-                                            class="inline-flex items-center justify-center w-8 h-8 text-sm text-gray-400 bg-transparent rounded-lg hover:bg-gray-200 hover:text-gray-900 ms-auto dark:hover:bg-gray-600 dark:hover:text-white"
-                                            data-modal-hide="barang">
-                                            <svg class="w-3 h-3" aria-hidden="true" xmlns="http://www.w3.org/2000/svg"
-                                                fill="none" viewBox="0 0 14 14">
-                                                <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
-                                                    stroke-width="2" d="m1 1 6 6m0 0 6 6M7 7l6-6M7 7l-6 6" />
-                                            </svg>
-                                            <span class="sr-only">Close modal</span>
-                                        </button>
-                                    </div>
-                                    <!-- Modal body -->
-
-                                    <form action="{{ route('data-admin-dan-staff.proses') }}" class="p-4 md:p-5"
-                                        method="POST">
-                                        @csrf
-                                        <div class="gap-2">
-                                            <div class="mb-2">
-                                                <label for="nama"
-                                                    class="block text-sm font-medium text-gray-700">Nama</label>
-                                                <input type="text" name="nama" id="nama" placeholder="Masukan Nama Lengkap"
-                                                    class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
-                                                @error('nama')
-                                                    <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
-                                                @enderror
-                                            </div>
-                                            <div class="mb-2">
-                                                <label for="nip"
-                                                    class="block text-sm font-medium text-gray-700">NIP</label>
-                                                <input type="number" name="nip" id="nip" placeholder="Masukan NIP"
-                                                    class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
-                                                @error('nip')
-                                                    <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
-                                                @enderror
-                                            </div>
-                                            <div class="mb-2">
-                                                <label for="username"
-                                                    class="block text-sm font-medium text-gray-700">Username</label>
-                                                <input type="text" name="username" id="username" placeholder="Masukan Username"
-                                                    class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
-                                                @error('username')
-                                                    <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
-                                                @enderror
-                                            </div>
-                                            <div class="mb-2">
-                                                <label for="password"
-                                                    class="block text-sm font-medium text-gray-700">Password</label>
-                                                <input type="password" name="password" id="password" placeholder="Masukan Password"
-                                                    class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
-                                                @error('password')
-                                                    <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
-                                                @enderror
-                                            </div>
-                                            <div class="mb-2">
-                                                <label for="role_id" class="block text-sm font-medium text-gray-700">Pilih
-                                                    Role</label>
-                                                <select name="role_id" id="role_id"
-                                                    class="mt-2 bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-green-500 focus:border-green-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-green-500 dark:focus:border-green-500">
-                                                    <option value="">- Pilih Role -</option>
-                                                    @foreach ($role as $role)
-                                                        <option value="{{ $role->id }}">{{ $role->nama_role }}</option>
-                                                    @endforeach
-                                                </select>
-                                                @error('role_id')
-                                                    <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
-                                                @enderror
-                                            </div>
-
-
-                                            <!-- Gambar -->
-                                            {{-- <div class="mb-2">
-                                                <label for="foto"
-                                                    class="block text-sm font-medium text-gray-700">Foto</label>
-                                                <input type="file" name="foto" id="foto"
-                                                    class="block w-full px-3 mt-1 border border-green-300 rounded-md shadow-sm focus:ring-green-500 focus:border-green-500 sm:text-sm">
-                                                @error('foto')
-                                                    <p class="mt-1 text-sm text-red-500">{{ $message }}</p>
-                                                @enderror
-                                            </div> --}}
-
-                                        </div>
-                                        <button type="submit"
-                                            class="text-white bg-green-500 mt-4 hover:bg-green-800 focus:ring-4 focus:outline-none focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center">Simpan</button>
-                                    </form>
-                                </div>
-                                </form>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="p-4 bg-white rounded-lg shadow-lg">
-                <form action="">
-
-                </form>
-                <div id="tableAdminDanStaff">
-                    @include('admin.pengguna.adminandstaff.table', ['users' => $user])
-                </div>
-
-                <div id="pageignitionLinks">
-                    {{ $user->links() }}
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <script>
-        function confirmDelete(id) {
-            Swal.fire({
-                title: 'Apakah Anda yakin?',
-                text: "Data yang dihapus tidak dapat dikembalikan!",
-                icon: 'warning',
-                showCancelButton: true,
-                confirmButtonColor: '#3085d6',
-                cancelButtonColor: '#d33',
-                confirmButtonText: 'Ya, hapus!',
-                cancelButtonText: 'Batal'
-            }).then((result) => {
-                if (result.isConfirmed) {
-                    document.getElementById('delete-form-' + id).submit();
-                }
-            })
-        }
-    </script>
-
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script>
-        $(document).ready(function() {
-            // Apply filter ketika button filter di klik
-            $('#applyFilter').on('click', function(e) {
-                e.preventDefault();
-                loadTable();
-            });
-
-            // Handle pagination link click event
-            $(document).on('click', '.pagination a', function(e) {
-                e.preventDefault();
-                var page = $(this).attr('href').split('page=')[1];
-                loadTable(page);
-            });
-
-            function loadTable(page = 1) {
-                $.ajax({
-                    url: "{{ route('data-admin-dan-staff') }}",
-                    method: "GET",
-                    data: {
-                        // name: $('#filterName').val(),
-                        // kategori_id: $('#filterKategori').val(),
-                        // kondisi: $('#filterKondisi').val(),
-                        // stock: $('#filterStock').val(),
-                        // satuan_id: $('#filterSatuan').val(),
-                        // page: page
-                    },
-                    success: function(response) {
-                        // Replace table and pagination links
-                        $('#tableAdminDanStaff').html($(response).find('#tableAdminDanStaff').html());
-                        $('#paginationLinks').html($(response).find('#paginationLinks').html());
-                    }
-                });
-            }
-        });
-        
-        $(document).ready(function() {
-            $('#data-admin-dan-staff').DataTable({
-                paging: false,
-                scrollCollapse: true,
-                scrollY: '300px'
-            });
-        });
-    </script>
-    
-@endsection
