--- resources/views/dosen/index.blade.php
@@ -1,37 +1,0 @@
-@extends('index')
-@section('content')
-    <div class="p-4 sm:ml-64">
-        <div class="rounded-lg mt-14 space-y-4">
-            @if (session('success'))
-                <script>
-                    Swal.fire({
-                        title: "Success",
-                        text: "{{ session('success') }}",
-                        icon: "success",
-                        confirmButtonColor: "#3085d6",
-                    });
-                </script>
-            @endif
-            
-        </div>
-    </div>
-
-    <script>
-        function confirmDelete(id) {
-            Swal.fire({
-                title: 'Apakah Anda yakin?',
-                text: "Data yang dihapus tidak dapat dikembalikan!",
-                icon: 'warning',
-                showCancelButton: true,
-                confirmButtonColor: '#3085d6',
-                cancelButtonColor: '#d33',
-                confirmButtonText: 'Ya, hapus!',
-                cancelButtonText: 'Batal'
-            }).then((result) => {
-                if (result.isConfirmed) {
-                    document.getElementById('delete-form-' + id).submit();
-                }
-            })
-        }
-    </script>
-@endsection
