+++ resources/views/pageAdmin/pengguna/adminandstaff/modal/edit.blade.php
@@ -1,0 +1,88 @@
+<div id="edit{{ $data->id }}" tabindex="-1" aria-hidden="true"
+    class="hidden overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 z-50 justify-center items-center w-full md:inset-0 h-[calc(100%-1rem)] max-h-full">
+    <div class="relative w-full max-w-2xl max-h-full p-4">
+        <!-- Modal content -->
+        <div class="relative bg-white rounded-lg shadow dark:bg-gray-700">
+            <!-- Modal header -->
+            <div
+                class="flex items-center justify-between p-4 border-b rounded-t md:p-5 dark:border-gray-600">
+                <h3 class="text-xl font-semibold text-gray-900 dark:text-white">
+                    Edit Admin dan Staff
+                </h3>
+                <button type="button"
+                    class="inline-flex items-center justify-center w-8 h-8 text-sm text-gray-400 bg-transparent rounded-lg hover:bg-gray-200 hover:text-gray-900 ms-auto dark:hover:bg-gray-600 dark:hover:text-white"
+                    data-modal-hide="edit{{ $data->id }}">
+                    <svg class="w-3 h-3" aria-hidden="true" xmlns="http://www.w3.org/2000/svg"
+                        fill="none" viewBox="0 0 14 14">
+                        <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
+                            stroke-width="2" d="m1 1 6 6m0 0 6 6M7 7l6-6M7 7l-6 6" />
+                    </svg>
+                    <span class="sr-only">Close modal</span>
+                </button>
+            </div>
+            <!-- Modal body -->
+            <div class="p-4">
+                <form action="{{ route('data-admin-dan-staff.edit', $data->id) }}" method="POST">
+                    @csrf
+                    @method('PUT')
+                    <div class="gap-2">
+                        <div class="mb-2">
+                            <label for="nama"
+                                class="block text-sm font-medium text-gray-700">Nama</label>
+                            <input type="text" name="nama" id="nama"
+                                value="{{ $data->nama }}"
+                                class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
+                            @error('nama')
+                                <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
+                            @enderror
+                        </div>
+                        <div class="mb-2">
+                            <label for="nip"
+                                class="block text-sm font-medium text-gray-700">NIP</label>
+                            <input type="number" name="nip" id="nip"
+                                value="{{ $data->nip }}"
+                                class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
+                            @error('nip')
+                                <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
+                            @enderror
+                        </div>
+                        <div class="mb-2">
+                            <label for="username"
+                                class="block text-sm font-medium text-gray-700">Username</label>
+                            <input type="text" name="username" id="username"
+                                value="{{ $data->username }}"
+                                class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
+                            @error('username')
+                                <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
+                            @enderror
+                        </div>
+                        <div class="mb-2">
+                            <label for="password"
+                                class="block text-sm font-medium text-gray-700">Password</label>
+                            <input type="password" name="password" id="password"
+                                class="block w-full mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
+                            @error('password')
+                                <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
+                            @enderror
+                        </div>
+                        <div class="mb-2">
+                            <label for="role_id"
+                                class="block text-sm font-medium text-gray-700">Pilih
+                                Role</label>
+                            <select name="role_id" id="role_id"
+                                class="block w-full p-2 mt-2 border-gray-300 rounded-lg shadow-sm focus:border-green-500 focus:ring focus:ring-green-500 focus:ring-opacity-50">
+                                <option value="{{ $data->role_id }}">{{ $data->role->nama_role }}
+                                </option>
+                            </select>
+                            @error('role_id')
+                                <p class="mt-1 text-sm text-red-600">{{ $message }}</p>
+                            @enderror
+                        </div>
+
+                        <button type="submit"
+                            class="text-white bg-green-500 mt-4 hover:bg-green-800 focus:ring-4 focus:outline-none focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">Simpan</button>
+                </form>
+            </div>
+        </div>
+    </div>
+</div>
