+++ resources/views/pageStaff/barang/actions.blade.php
