+++ resources/views/pageStaff/ruangan/index.blade.php
